Akıllı Hava Durumu/requirements.txt
PyQt6
matplotlib
requests
Pillow